[shardformer] refactored the user api (#3828)
* [shardformer] refactored the user api

* polish code
